backend/pyproject.toml
[tool.poetry]
name = "agentos-backend"
version = "0.1.0"
description = "AgentOS Core Backend"
authors = ["Your Name <youremail@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
motor = "^3.4.0"
celery = "^5.3.6"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.2.0"
loguru = "^0.7.2"
openai = "^1.16.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
httpx = "^0.27.0"
slowapi = "^0.1.9"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.27.0"
pytest-mock = "^3.10.0"
mongomock-motor = "^0.0.2a2"
testcontainers = { extras = ["mongodb", "redis"], version = "^4.4.0" }
pytest-dotenv = "^0.5.2"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"